cleanup(core): split project-configuration-utils into focused modules
Decompose the monolithic project-configuration-utils.ts (~1400 lines)
into focused modules with a clear dependency DAG:

- target-merging.ts: target-level merge utilities (leaf module)
- target-normalization.ts: normalization, validation, target defaults
- project-nodes-manager.ts: project-level merge and read operations

Split the corresponding spec file into sibling spec files for each
new module. All 143 tests continue to pass.
