+++ packages/nx/src/devkit-internals.ts
@@ -11,11 +11,9 @@ export {
 export { readNxJson as readNxJsonFromDisk } from './config/nx-json';
 export { calculateDefaultProjectName } from './config/calculate-default-project-name';
 export { retrieveProjectConfigurationsWithAngularProjects } from './project-graph/utils/retrieve-workspace-files';
-export { mergeTargetConfigurations } from './project-graph/utils/project-configuration-utils';
-export {
-  readProjectConfigurationsFromRootMap,
-  findMatchingConfigFiles,
-} from './project-graph/utils/project-configuration-utils';
+export { mergeTargetConfigurations } from './project-graph/utils/project-configuration/target-merging';
+export { readProjectConfigurationsFromRootMap } from './project-graph/utils/project-configuration/project-nodes-manager';
+export { findMatchingConfigFiles } from './project-graph/utils/project-configuration-utils';
 export { getIgnoreObjectForTree } from './utils/ignore';
 export { splitTarget } from './utils/split-target';
 export { combineOptionsForExecutor } from './utils/params';
