+++ packages/nx/src/project-graph/utils/project-configuration-utils.ts
@@ -1,49 +1,28 @@
-import { NxJsonConfiguration, TargetDefaults } from '../../config/nx-json';
+import { NxJsonConfiguration } from '../../config/nx-json';
 import { ProjectGraphExternalNode } from '../../config/project-graph';
-import {
-  ProjectConfiguration,
-  ProjectMetadata,
-  TargetConfiguration,
-  TargetMetadata,
-} from '../../config/workspace-json-project-json';
-import { readJsonFile } from '../../utils/fileutils';
-import { NX_PREFIX } from '../../utils/logger';
+import { ProjectConfiguration } from '../../config/workspace-json-project-json';
 import { workspaceRoot } from '../../utils/workspace-root';
-import { ProjectNameInNodePropsManager } from './project-configuration/name-substitution-manager';
 import {
-  recordSourceMapKeysByIndex,
-  targetConfigurationsSourceMapKey,
-  targetOptionSourceMapKey,
-  targetSourceMapKey,
-} from './project-configuration/source-maps';
+  createRootMap,
+  ProjectNodesManager,
+} from './project-configuration/project-nodes-manager';
+import { validateAndNormalizeProjectRootMap } from './project-configuration/target-normalization';
 
 import { minimatch } from 'minimatch';
-import { existsSync } from 'node:fs';
-import { join } from 'path';
 import { performance } from 'perf_hooks';
 
-import {
-  getExecutorInformation,
-  parseExecutor,
-} from '../../command-line/run/executor-utils';
-import { toProjectName } from '../../config/to-project-name';
 import { DelayedSpinner } from '../../utils/delayed-spinner';
-import { isGlobPattern } from '../../utils/globs';
 import {
   AggregateCreateNodesError,
   formatAggregateCreateNodesError,
   isAggregateCreateNodesError,
   isMultipleProjectsWithSameNameError,
   isProjectsWithNoNameError,
-  isProjectWithExistingNameError,
-  isProjectWithNoNameError,
   isWorkspaceValidityError,
   MergeNodesError,
   MultipleProjectsWithSameNameError,
   ProjectConfigurationsError,
   ProjectsWithNoNameError,
-  ProjectWithExistingNameError,
-  ProjectWithNoNameError,
   WorkspaceValidityError,
 } from '../error-types';
 import type { LoadedNxPlugin } from '../plugins/loaded-nx-plugin';
@@ -53,289 +32,27 @@ import type {
   ConfigurationSourceMaps,
   SourceInformation,
 } from './project-configuration/source-maps';
+
+// Re-export moved functions for consumers that haven't updated their imports yet.
+// Also re-export types that were previously exported from this file.
 export type { ConfigurationSourceMaps, SourceInformation };
-
-export function mergeProjectConfigurationIntoRootMap(
-  projectRootMap: Record<string, ProjectConfiguration>,
-  project: ProjectConfiguration,
-  configurationSourceMaps?: ConfigurationSourceMaps,
-  sourceInformation?: SourceInformation,
-  // This function is used when reading project configuration
-  // in generators, where we don't want to do this.
-  skipTargetNormalization?: boolean
-): {
-  previousName?: string;
-} {
-  project.root = project.root === '' ? '.' : project.root;
-  if (configurationSourceMaps && !configurationSourceMaps[project.root]) {
-    configurationSourceMaps[project.root] = {};
-  }
-  const sourceMap = configurationSourceMaps?.[project.root];
-
-  let matchingProject = projectRootMap[project.root];
-
-  if (!matchingProject) {
-    projectRootMap[project.root] = {
-      root: project.root,
-    };
-    matchingProject = projectRootMap[project.root];
-    if (sourceMap) {
-      sourceMap[`root`] = sourceInformation;
-    }
-  }
-
-  // This handles top level properties that are overwritten.
-  // e.g. `srcRoot`, `projectType`, or other fields that shouldn't be extended
-  // Note: `name` is set specifically here to keep it from changing. The name is
-  // always determined by the first inference plugin to ID a project, unless it has
-  // a project.json in which case it was already updated above.
-  const updatedProjectConfiguration = {
-    ...matchingProject,
-  };
-
-  for (const k in project) {
-    if (
-      ![
-        'tags',
-        'implicitDependencies',
-        'generators',
-        'targets',
-        'metadata',
-        'namedInputs',
-      ].includes(k)
-    ) {
-      updatedProjectConfiguration[k] = project[k];
-      if (sourceMap) {
-        sourceMap[`${k}`] = sourceInformation;
-      }
-    }
-  }
-
-  // The next blocks handle properties that should be themselves merged (e.g. targets, tags, and implicit dependencies)
-  if (project.tags) {
-    updatedProjectConfiguration.tags = Array.from(
-      new Set((matchingProject.tags ?? []).concat(project.tags))
-    );
-
-    if (sourceMap) {
-      sourceMap['tags'] ??= sourceInformation;
-      project.tags.forEach((tag) => {
-        sourceMap[`tags.${tag}`] = sourceInformation;
-      });
-    }
-  }
-
-  if (project.implicitDependencies) {
-    updatedProjectConfiguration.implicitDependencies = (
-      matchingProject.implicitDependencies ?? []
-    ).concat(project.implicitDependencies);
-
-    if (sourceMap) {
-      sourceMap['implicitDependencies'] ??= sourceInformation;
-      project.implicitDependencies.forEach((implicitDependency) => {
-        sourceMap[`implicitDependencies.${implicitDependency}`] =
-          sourceInformation;
-      });
-    }
-  }
-
-  if (project.generators) {
-    // Start with generators config in new project.
-    updatedProjectConfiguration.generators = { ...project.generators };
-
-    if (sourceMap) {
-      sourceMap['generators'] ??= sourceInformation;
-      for (const generator in project.generators) {
-        sourceMap[`generators.${generator}`] = sourceInformation;
-        for (const property in project.generators[generator]) {
-          sourceMap[`generators.${generator}.${property}`] = sourceInformation;
-        }
-      }
-    }
-
-    if (matchingProject.generators) {
-      // For each generator that was already defined, shallow merge the options.
-      // Project contains the new info, so it has higher priority.
-      for (const generator in matchingProject.generators) {
-        updatedProjectConfiguration.generators[generator] = {
-          ...matchingProject.generators[generator],
-          ...project.generators[generator],
-        };
-      }
-    }
-  }
-
-  if (project.namedInputs) {
-    updatedProjectConfiguration.namedInputs = {
-      ...matchingProject.namedInputs,
-      ...project.namedInputs,
-    };
-
-    if (sourceMap) {
-      sourceMap['namedInputs'] ??= sourceInformation;
-      for (const namedInput in project.namedInputs) {
-        sourceMap[`namedInputs.${namedInput}`] = sourceInformation;
-      }
-    }
-  }
-
-  if (project.metadata) {
-    updatedProjectConfiguration.metadata = mergeMetadata(
-      sourceMap,
-      sourceInformation,
-      'metadata',
-      project.metadata,
-      matchingProject.metadata
-    );
-  }
-
-  if (project.targets) {
-    // We merge the targets with special handling, so clear this back to the
-    // targets as defined originally before merging.
-    updatedProjectConfiguration.targets = matchingProject?.targets ?? {};
-    if (sourceMap) {
-      sourceMap['targets'] ??= sourceInformation;
-    }
-
-    // For each target defined in the new config
-    for (const targetName in project.targets) {
-      // Always set source map info for the target, but don't overwrite info already there
-      // if augmenting an existing target.
-
-      const target = project.targets?.[targetName];
-
-      if (sourceMap) {
-        sourceMap[targetSourceMapKey(targetName)] = sourceInformation;
-      }
-
-      const normalizedTarget = skipTargetNormalization
-        ? target
-        : resolveCommandSyntacticSugar(target, project.root);
-
-      let matchingTargets = [];
-      if (isGlobPattern(targetName)) {
-        // find all targets matching the glob pattern
-        // this will map atomized targets to the glob pattern same as it does for targetDefaults
-        matchingTargets = Object.keys(
-          updatedProjectConfiguration.targets
-        ).filter((key) => minimatch(key, targetName));
-      }
-      // If no matching targets were found, we can assume that the target name is not (meant to be) a glob pattern
-      if (!matchingTargets.length) {
-        matchingTargets = [targetName];
-      }
-
-      for (const matchingTargetName of matchingTargets) {
-        const mergedTarget = mergeTargetConfigurations(
-          normalizedTarget,
-          matchingProject.targets?.[matchingTargetName],
-          sourceMap,
-          sourceInformation,
-          `targets.${matchingTargetName}`
-        );
-
-        updatedProjectConfiguration.targets[matchingTargetName] = mergedTarget;
-      }
-    }
-  }
-
-  projectRootMap[updatedProjectConfiguration.root] =
-    updatedProjectConfiguration;
-
-  const previousName =
-    matchingProject?.name &&
-    project.name &&
-    matchingProject.name !== project.name
-      ? matchingProject.name
-      : undefined;
-
-  return { previousName };
-}
-
-export function mergeMetadata<T = ProjectMetadata | TargetMetadata>(
-  sourceMap: Record<string, [file: string, plugin: string]>,
-  sourceInformation: [file: string, plugin: string],
-  baseSourceMapPath: string,
-  metadata: T,
-  matchingMetadata?: T
-): T {
-  const result: T = {
-    ...(matchingMetadata ?? ({} as T)),
-  };
-  for (const [metadataKey, value] of Object.entries(metadata)) {
-    const existingValue = matchingMetadata?.[metadataKey];
-
-    if (Array.isArray(value) && Array.isArray(existingValue)) {
-      const startIndex = result[metadataKey].length;
-      result[metadataKey].push(...value);
-      if (sourceMap) {
-        recordSourceMapKeysByIndex(
-          sourceMap,
-          `${baseSourceMapPath}.${metadataKey}`,
-          result[metadataKey],
-          sourceInformation,
-          startIndex
-        );
-      }
-    } else if (Array.isArray(value) && existingValue === undefined) {
-      result[metadataKey] ??= value;
-      if (sourceMap) {
-        sourceMap[`${baseSourceMapPath}.${metadataKey}`] = sourceInformation;
-        recordSourceMapKeysByIndex(
-          sourceMap,
-          `${baseSourceMapPath}.${metadataKey}`,
-          value,
-          sourceInformation
-        );
-      }
-    } else if (typeof value === 'object' && typeof existingValue === 'object') {
-      for (const key in value) {
-        const existingValue = matchingMetadata?.[metadataKey]?.[key];
-
-        if (Array.isArray(value[key]) && Array.isArray(existingValue)) {
-          const startIndex = result[metadataKey][key].length;
-          result[metadataKey][key].push(...value[key]);
-          if (sourceMap) {
-            recordSourceMapKeysByIndex(
-              sourceMap,
-              `${baseSourceMapPath}.${metadataKey}.${key}`,
-              result[metadataKey][key],
-              sourceInformation,
-              startIndex
-            );
-          }
-        } else {
-          result[metadataKey][key] = value[key];
-          if (sourceMap) {
-            sourceMap[`${baseSourceMapPath}.${metadataKey}`] =
-              sourceInformation;
-          }
-        }
-      }
-    } else {
-      result[metadataKey] = value;
-      if (sourceMap) {
-        sourceMap[`${baseSourceMapPath}.${metadataKey}`] = sourceInformation;
-
-        if (typeof value === 'object') {
-          for (const k in value) {
-            sourceMap[`${baseSourceMapPath}.${metadataKey}.${k}`] =
-              sourceInformation;
-            if (Array.isArray(value[k])) {
-              recordSourceMapKeysByIndex(
-                sourceMap,
-                `${baseSourceMapPath}.${metadataKey}.${k}`,
-                value[k],
-                sourceInformation
-              );
-            }
-          }
-        }
-      }
-    }
-  }
-  return result;
-}
+export {
+  mergeProjectConfigurationIntoRootMap,
+  readProjectConfigurationsFromRootMap,
+  validateProject,
+} from './project-configuration/project-nodes-manager';
+export {
+  mergeMetadata,
+  mergeTargetConfigurations,
+  isCompatibleTarget,
+  mergeTargetDefaultWithTargetDefinition,
+  resolveNxTokensInOptions,
+  readTargetDefaultsForTarget,
+} from './project-configuration/target-merging';
+export {
+  normalizeTarget,
+  validateAndNormalizeProjectRootMap,
+} from './project-configuration/target-normalization';
 
 export type ConfigurationResult = {
   /**
@@ -513,7 +230,7 @@ export async function createProjectConfigurationsWithPlugins(
   });
 }
 
-function mergeCreateNodesResults(
+export function mergeCreateNodesResults(
   results: (readonly [
     plugin: string,
     file: string,
@@ -531,44 +248,72 @@ function mergeCreateNodesResults(
   )[]
 ) {
   performance.mark('createNodes:merge - start');
-  const projectRootMap: Record<string, ProjectConfiguration> = {};
+  const nodesManager = new ProjectNodesManager();
   const externalNodes: Record<string, ProjectGraphExternalNode> = {};
-  const projectNameManager = new ProjectNameInNodePropsManager();
   const configurationSourceMaps: Record<
     string,
     Record<string, SourceInformation>
   > = {};
 
-  for (const result of results.flat()) {
-    const [pluginName, file, nodes, pluginIndex] = result;
+  // Process each plugin's results in two phases:
+  //   Phase 1: Merge all projects from this plugin into rootMap/nameMap
+  //   Phase 2: Register substitutors for this plugin's results
+  //
+  // Per-plugin batching ensures that:
+  //  - All same-plugin projects are in the nameMap before substitutor
+  //    registration (fixes cross-file references like kafka-stream)
+  //  - Later-plugin renames haven't occurred yet, so dependsOn strings
+  //    that reference old names can still be resolved via the nameMap
+  for (const pluginResults of results) {
+    // Phase 1: Merge all projects from this plugin batch
+    for (const result of pluginResults) {
+      const [pluginName, file, nodes, pluginIndex] = result;
 
-    const { projects: projectNodes, externalNodes: pluginExternalNodes } =
-      nodes;
+      const { projects: projectNodes, externalNodes: pluginExternalNodes } =
+        nodes;
 
-    const sourceInfo: SourceInformation = [file, pluginName];
+      const sourceInfo: SourceInformation = [file, pluginName];
 
-    for (const root in projectNodes) {
-      // Handles `{projects: {'libs/foo': undefined}}`.
-      if (!projectNodes[root]) {
-        continue;
+      for (const root in projectNodes) {
+        // Handles `{projects: {'libs/foo': undefined}}`.
+        if (!projectNodes[root]) {
+          continue;
+        }
+        const project = {
+          root: root,
+          ...projectNodes[root],
+        };
+
+        try {
+          nodesManager.mergeProjectNode(
+            project,
+            configurationSourceMaps,
+            sourceInfo
+          );
+        } catch (error) {
+          errors.push(
+            new MergeNodesError({
+              file,
+              pluginName,
+              error,
+              pluginIndex,
+            })
+          );
+        }
       }
-      const project = {
-        root: root,
-        ...projectNodes[root],
-      };
+
+      Object.assign(externalNodes, pluginExternalNodes);
+    }
+
+    // Phase 2: Register substitutors for this plugin batch. The nameMap
+    // now contains all projects from this plugin (and all prior plugins)
+    // so splitTargetFromConfigurations can resolve colon-delimited strings.
+    for (const result of pluginResults) {
+      const [pluginName, file, nodes, pluginIndex] = result;
+      const { projects: projectNodes } = nodes;
 
       try {
-        const { previousName } = mergeProjectConfigurationIntoRootMap(
-          projectRootMap,
-          project,
-          configurationSourceMaps,
-          sourceInfo
-        );
-        // If this project's name changed, record the rename so substitutors
-        // registered for the old name can fire during applySubstitutions.
-        if (previousName) {
-          projectNameManager.markDirty(root, previousName);
-        }
+        nodesManager.registerSubstitutors(projectNodes);
       } catch (error) {
         errors.push(
           new MergeNodesError({
@@ -580,27 +325,12 @@ function mergeCreateNodesResults(
         );
       }
     }
-    // Register substitutors for any project-name references in this result's
-    // targets. Substitutors are keyed by the referenced name (not by root),
-    // so registration requires no lookup and is safe regardless of whether
-    // the referenced project has been processed yet.
-    try {
-      projectNameManager.registerSubstitutorsForNodeResults(projectNodes);
-    } catch (error) {
-      errors.push(
-        new MergeNodesError({
-          file,
-          pluginName,
-          error,
-          pluginIndex,
-        })
-      );
-    }
-    Object.assign(externalNodes, pluginExternalNodes);
   }
 
+  const projectRootMap = nodesManager.getRootMap();
+
   try {
-    projectNameManager.applySubstitutions(projectRootMap);
+    nodesManager.applySubstitutions();
     validateAndNormalizeProjectRootMap(
       workspaceRoot,
       projectRootMap,
@@ -608,7 +338,6 @@ function mergeCreateNodesResults(
       configurationSourceMaps
     );
   } catch (error) {
-    const unknownErrors: Error[] = [];
     let _errors = error instanceof AggregateError ? error.errors : [error];
     for (const e of _errors) {
       if (
@@ -712,696 +441,4 @@ export function findMatchingConfigFiles(
   }
 
   return matchingConfigFiles;
-}
-
-export function readProjectConfigurationsFromRootMap(
-  projectRootMap: Record<string, ProjectConfiguration>
-) {
-  const projects: Record<string, ProjectConfiguration> = {};
-  // If there are projects that have the same name, that is an error.
-  // This object tracks name -> (all roots of projects with that name)
-  // to provide better error messaging.
-  const conflicts = new Map<string, string[]>();
-  const projectRootsWithNoName: string[] = [];
-
-  for (const root in projectRootMap) {
-    const project = projectRootMap[root];
-    // We're setting `// targets` as a comment `targets` is empty due to Project Crystal.
-    // Strip it before returning configuration for usage.
-    if (project['// targets']) delete project['// targets'];
-
-    try {
-      validateProject(project, projects);
-      projects[project.name] = project;
-    } catch (e) {
-      if (isProjectWithNoNameError(e)) {
-        projectRootsWithNoName.push(e.projectRoot);
-      } else if (isProjectWithExistingNameError(e)) {
-        const rootErrors = conflicts.get(e.projectName) ?? [
-          projects[e.projectName].root,
-        ];
-        rootErrors.push(e.projectRoot);
-        conflicts.set(e.projectName, rootErrors);
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  if (conflicts.size > 0) {
-    throw new MultipleProjectsWithSameNameError(conflicts, projects);
-  }
-  if (projectRootsWithNoName.length > 0) {
-    throw new ProjectsWithNoNameError(projectRootsWithNoName, projects);
-  }
-  return projects;
-}
-
-function validateAndNormalizeProjectRootMap(
-  workspaceRoot: string,
-  projectRootMap: Record<string, ProjectConfiguration>,
-  nxJsonConfiguration: NxJsonConfiguration,
-  sourceMaps: ConfigurationSourceMaps = {}
-) {
-  // Name -> Project, used to validate that all projects have unique names
-  const projects: Record<string, ProjectConfiguration> = {};
-  // If there are projects that have the same name, that is an error.
-  // This object tracks name -> (all roots of projects with that name)
-  // to provide better error messaging.
-  const conflicts = new Map<string, string[]>();
-  const projectRootsWithNoName: string[] = [];
-  const validityErrors: WorkspaceValidityError[] = [];
-
-  for (const root in projectRootMap) {
-    const project = projectRootMap[root];
-    // We're setting `// targets` as a comment `targets` is empty due to Project Crystal.
-    // Strip it before returning configuration for usage.
-    if (project['// targets']) delete project['// targets'];
-
-    // We initially did this in the project.json plugin, but
-    // that resulted in project.json files without names causing
-    // the resulting project to change names from earlier plugins...
-    if (
-      !project.name &&
-      existsSync(join(workspaceRoot, project.root, 'project.json'))
-    ) {
-      project.name = toProjectName(join(root, 'project.json'));
-    }
-
-    try {
-      validateProject(project, projects);
-      projects[project.name] = project;
-    } catch (e) {
-      if (isProjectWithNoNameError(e)) {
-        projectRootsWithNoName.push(e.projectRoot);
-      } else if (isProjectWithExistingNameError(e)) {
-        const rootErrors = conflicts.get(e.projectName) ?? [
-          projects[e.projectName].root,
-        ];
-        rootErrors.push(e.projectRoot);
-        conflicts.set(e.projectName, rootErrors);
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  for (const root in projectRootMap) {
-    const project = projectRootMap[root];
-    try {
-      normalizeTargets(
-        project,
-        sourceMaps,
-        nxJsonConfiguration,
-        workspaceRoot,
-        projects
-      );
-    } catch (e) {
-      if (e instanceof WorkspaceValidityError) {
-        validityErrors.push(e);
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  const errors: Error[] = [];
-
-  if (conflicts.size > 0) {
-    errors.push(new MultipleProjectsWithSameNameError(conflicts, projects));
-  }
-  if (projectRootsWithNoName.length > 0) {
-    errors.push(new ProjectsWithNoNameError(projectRootsWithNoName, projects));
-  }
-  if (validityErrors.length > 0) {
-    errors.push(...validityErrors);
-  }
-  if (errors.length > 0) {
-    throw new AggregateError(errors);
-  }
-  return projectRootMap;
-}
-
-function normalizeTargets(
-  project: ProjectConfiguration,
-  sourceMaps: ConfigurationSourceMaps,
-  nxJsonConfiguration: NxJsonConfiguration,
-  workspaceRoot: string,
-  /**
-   * Project configurations keyed by project name
-   */
-  projects: Record<string, ProjectConfiguration>
-) {
-  const targetErrorMessage: string[] = [];
-
-  for (const targetName in project.targets) {
-    project.targets[targetName] = normalizeTarget(
-      project.targets[targetName],
-      project,
-      workspaceRoot,
-      projects,
-      [project.root, targetName].join(':')
-    );
-
-    const projectSourceMaps = sourceMaps[project.root];
-
-    const targetConfig = project.targets[targetName];
-    const targetDefaults = deepClone(
-      readTargetDefaultsForTarget(
-        targetName,
-        nxJsonConfiguration.targetDefaults,
-        targetConfig.executor
-      )
-    );
-
-    // We only apply defaults if they exist
-    if (targetDefaults && isCompatibleTarget(targetConfig, targetDefaults)) {
-      project.targets[targetName] = mergeTargetDefaultWithTargetDefinition(
-        targetName,
-        project,
-        normalizeTarget(
-          targetDefaults,
-          project,
-          workspaceRoot,
-          projects,
-          ['nx.json[targetDefaults]', targetName].join(':')
-        ),
-        projectSourceMaps
-      );
-    }
-
-    const target = project.targets[targetName];
-
-    if (
-      // If the target has no executor or command, it doesn't do anything
-      !target.executor &&
-      !target.command
-    ) {
-      // But it may have dependencies that do something
-      if (target.dependsOn && target.dependsOn.length > 0) {
-        target.executor = 'nx:noop';
-      } else {
-        // If it does nothing, and has no depenencies,
-        // we can remove it.
-        delete project.targets[targetName];
-      }
-    }
-
-    if (target.cache && target.continuous) {
-      targetErrorMessage.push(
-        `- "${targetName}" has both "cache" and "continuous" set to true. Continuous targets cannot be cached. Please remove the "cache" property.`
-      );
-    }
-  }
-  if (targetErrorMessage.length > 0) {
-    targetErrorMessage.unshift(
-      `Errors detected in targets of project "${project.name}":`
-    );
-    throw new WorkspaceValidityError(targetErrorMessage.join('\n'));
-  }
-}
-
-export function validateProject(
-  project: ProjectConfiguration,
-  // name -> project
-  knownProjects: Record<string, ProjectConfiguration>
-) {
-  if (!project.name) {
-    try {
-      const { name } = readJsonFile(join(project.root, 'package.json'));
-      if (!name) {
-        throw new Error(`Project at ${project.root} has no name provided.`);
-      }
-      project.name = name;
-    } catch {
-      throw new ProjectWithNoNameError(project.root);
-    }
-  } else if (
-    knownProjects[project.name] &&
-    knownProjects[project.name].root !== project.root
-  ) {
-    throw new ProjectWithExistingNameError(project.name, project.root);
-  }
-}
-
-function targetDefaultShouldBeApplied(
-  key: string,
-  sourceMap: Record<string, SourceInformation>
-) {
-  const sourceInfo = sourceMap[key];
-  if (!sourceInfo) {
-    return true;
-  }
-  // The defined value of the target is from a plugin that
-  // isn't part of Nx's core plugins, so target defaults are
-  // applied on top of it.
-  const [, plugin] = sourceInfo;
-  return !plugin?.startsWith('nx/');
-}
-
-function deepClone(obj) {
-  return structuredClone(obj);
-}
-
-export function mergeTargetDefaultWithTargetDefinition(
-  targetName: string,
-  project: ProjectConfiguration,
-  targetDefault: Partial<TargetConfiguration>,
-  sourceMap: Record<string, SourceInformation>
-): TargetConfiguration {
-  const targetDefinition = project.targets[targetName] ?? {};
-  const result = deepClone(targetDefinition);
-
-  for (const key in targetDefault) {
-    switch (key) {
-      case 'options': {
-        const normalizedDefaults = resolveNxTokensInOptions(
-          targetDefault.options,
-          project,
-          targetName
-        );
-        for (const optionKey in normalizedDefaults) {
-          const sourceMapKey = targetOptionSourceMapKey(targetName, optionKey);
-          if (
-            targetDefinition.options[optionKey] === undefined ||
-            targetDefaultShouldBeApplied(sourceMapKey, sourceMap)
-          ) {
-            result.options[optionKey] = targetDefault.options[optionKey];
-            sourceMap[sourceMapKey] = ['nx.json', 'nx/target-defaults'];
-          }
-        }
-        break;
-      }
-      case 'configurations': {
-        if (!result.configurations) {
-          result.configurations = {};
-          sourceMap[targetConfigurationsSourceMapKey(targetName)] = [
-            'nx.json',
-            'nx/target-defaults',
-          ];
-        }
-        for (const configuration in targetDefault.configurations) {
-          if (!result.configurations[configuration]) {
-            result.configurations[configuration] = {};
-            sourceMap[
-              targetConfigurationsSourceMapKey(targetName, configuration)
-            ] = ['nx.json', 'nx/target-defaults'];
-          }
-          const normalizedConfigurationDefaults = resolveNxTokensInOptions(
-            targetDefault.configurations[configuration],
-            project,
-            targetName
-          );
-          for (const configurationKey in normalizedConfigurationDefaults) {
-            const sourceMapKey = targetConfigurationsSourceMapKey(
-              targetName,
-              configuration,
-              configurationKey
-            );
-            if (
-              targetDefinition.configurations?.[configuration]?.[
-                configurationKey
-              ] === undefined ||
-              targetDefaultShouldBeApplied(sourceMapKey, sourceMap)
-            ) {
-              result.configurations[configuration][configurationKey] =
-                targetDefault.configurations[configuration][configurationKey];
-              sourceMap[sourceMapKey] = ['nx.json', 'nx/target-defaults'];
-            }
-          }
-        }
-        break;
-      }
-      default: {
-        const sourceMapKey = `targets.${targetName}.${key}`;
-        if (
-          targetDefinition[key] === undefined ||
-          targetDefaultShouldBeApplied(sourceMapKey, sourceMap)
-        ) {
-          result[key] = targetDefault[key];
-          sourceMap[sourceMapKey] = ['nx.json', 'nx/target-defaults'];
-        }
-        break;
-      }
-    }
-  }
-  return result;
-}
-
-/**
- * Merges two targets.
- *
- * Most properties from `target` will overwrite any properties from `baseTarget`.
- * Options and configurations are treated differently - they are merged together if the executor definition is compatible.
- *
- * @param target The target definition with higher priority
- * @param baseTarget The target definition that should be overwritten. Can be undefined, in which case the target is returned as-is.
- * @param projectConfigSourceMap The source map to be filled with metadata about where each property came from
- * @param sourceInformation The metadata about where the new target was defined
- * @param targetIdentifier The identifier for the target to merge, used for source map
- * @returns A merged target configuration
- */
-export function mergeTargetConfigurations(
-  target: TargetConfiguration,
-  baseTarget?: TargetConfiguration,
-  projectConfigSourceMap?: Record<string, SourceInformation>,
-  sourceInformation?: SourceInformation,
-  targetIdentifier?: string
-): TargetConfiguration {
-  const {
-    configurations: defaultConfigurations,
-    options: defaultOptions,
-    ...baseTargetProperties
-  } = baseTarget ?? {};
-
-  // Target is "compatible", e.g. executor is defined only once or is the same
-  // in both places. This means that it is likely safe to merge
-  const isCompatible = isCompatibleTarget(baseTarget ?? {}, target);
-
-  if (!isCompatible && projectConfigSourceMap) {
-    // if the target is not compatible, we will simply override the options
-    // we have to delete old entries from the source map
-    for (const key in projectConfigSourceMap) {
-      if (key.startsWith(`${targetIdentifier}`)) {
-        delete projectConfigSourceMap[key];
-      }
-    }
-  }
-
-  // merge top level properties if they're compatible
-  const result = {
-    ...(isCompatible ? baseTargetProperties : {}),
-    ...target,
-  };
-
-  // record top level properties in source map
-  if (projectConfigSourceMap) {
-    projectConfigSourceMap[targetIdentifier] = sourceInformation;
-
-    // record root level target properties to source map
-    for (const targetProperty in target) {
-      const targetPropertyId = `${targetIdentifier}.${targetProperty}`;
-      projectConfigSourceMap[targetPropertyId] = sourceInformation;
-    }
-  }
-
-  // merge options if there are any
-  // if the targets aren't compatible, we simply discard the old options during the merge
-  if (target.options || defaultOptions) {
-    result.options = mergeOptions(
-      target.options,
-      isCompatible ? defaultOptions : undefined,
-      projectConfigSourceMap,
-      sourceInformation,
-      targetIdentifier
-    );
-  }
-
-  // merge configurations if there are any
-  // if the targets aren't compatible, we simply discard the old configurations during the merge
-  if (target.configurations || defaultConfigurations) {
-    result.configurations = mergeConfigurations(
-      target.configurations,
-      isCompatible ? defaultConfigurations : undefined,
-      projectConfigSourceMap,
-      sourceInformation,
-      targetIdentifier
-    );
-  }
-
-  if (target.metadata) {
-    result.metadata = mergeMetadata(
-      projectConfigSourceMap,
-      sourceInformation,
-      `${targetIdentifier}.metadata`,
-      target.metadata,
-      baseTarget?.metadata
-    );
-  }
-
-  return result as TargetConfiguration;
-}
-
-/**
- * Checks if targets options are compatible - used when merging configurations
- * to avoid merging options for @nx/js:tsc into something like @nx/webpack:webpack.
- *
- * If the executors are both specified and don't match, the options aren't considered
- * "compatible" and shouldn't be merged.
- */
-export function isCompatibleTarget(
-  a: TargetConfiguration,
-  b: TargetConfiguration
-) {
-  const oneHasNoExecutor = !a.executor || !b.executor;
-  const bothHaveSameExecutor = a.executor === b.executor;
-
-  if (oneHasNoExecutor) return true;
-  if (!bothHaveSameExecutor) return false;
-
-  const isRunCommands = a.executor === 'nx:run-commands';
-  if (isRunCommands) {
-    const aCommand = a.options?.command ?? a.options?.commands?.join(' && ');
-    const bCommand = b.options?.command ?? b.options?.commands?.join(' && ');
-
-    const oneHasNoCommand = !aCommand || !bCommand;
-    const hasSameCommand = aCommand === bCommand;
-
-    return oneHasNoCommand || hasSameCommand;
-  }
-
-  const isRunScript = a.executor === 'nx:run-script';
-  if (isRunScript) {
-    const aScript = a.options?.script;
-    const bScript = b.options?.script;
-
-    const oneHasNoScript = !aScript || !bScript;
-    const hasSameScript = aScript === bScript;
-
-    return oneHasNoScript || hasSameScript;
-  }
-
-  return true;
-}
-
-function mergeConfigurations<T extends Object>(
-  newConfigurations: Record<string, T> | undefined,
-  baseConfigurations: Record<string, T> | undefined,
-  projectConfigSourceMap?: Record<string, SourceInformation>,
-  sourceInformation?: SourceInformation,
-  targetIdentifier?: string
-): Record<string, T> | undefined {
-  const mergedConfigurations = {};
-
-  const configurations = new Set([
-    ...Object.keys(baseConfigurations ?? {}),
-    ...Object.keys(newConfigurations ?? {}),
-  ]);
-  for (const configuration of configurations) {
-    mergedConfigurations[configuration] = {
-      ...(baseConfigurations?.[configuration] ?? {}),
-      ...(newConfigurations?.[configuration] ?? {}),
-    };
-  }
-
-  // record new configurations & configuration properties in source map
-  if (projectConfigSourceMap) {
-    for (const newConfiguration in newConfigurations) {
-      projectConfigSourceMap[
-        `${targetIdentifier}.configurations.${newConfiguration}`
-      ] = sourceInformation;
-      for (const configurationProperty in newConfigurations[newConfiguration]) {
-        projectConfigSourceMap[
-          `${targetIdentifier}.configurations.${newConfiguration}.${configurationProperty}`
-        ] = sourceInformation;
-      }
-    }
-  }
-
-  return mergedConfigurations;
-}
-
-function mergeOptions(
-  newOptions: Record<string, any> | undefined,
-  baseOptions: Record<string, any> | undefined,
-  projectConfigSourceMap?: Record<string, SourceInformation>,
-  sourceInformation?: SourceInformation,
-  targetIdentifier?: string
-): Record<string, any> | undefined {
-  const mergedOptions = {
-    ...(baseOptions ?? {}),
-    ...(newOptions ?? {}),
-  };
-
-  // record new options & option properties in source map
-  if (projectConfigSourceMap) {
-    for (const newOption in newOptions) {
-      projectConfigSourceMap[`${targetIdentifier}.options.${newOption}`] =
-        sourceInformation;
-    }
-  }
-
-  return mergedOptions;
-}
-
-export function resolveNxTokensInOptions<T extends Object | Array<unknown>>(
-  object: T,
-  project: ProjectConfiguration,
-  key: string
-): T {
-  const result: T = Array.isArray(object) ? ([...object] as T) : { ...object };
-  for (let [opt, value] of Object.entries(object ?? {})) {
-    if (typeof value === 'string') {
-      const workspaceRootMatch = /^(\{workspaceRoot\}\/?)/.exec(value);
-      if (workspaceRootMatch?.length) {
-        value = value.replace(workspaceRootMatch[0], '');
-      }
-      if (value.includes('{workspaceRoot}')) {
-        throw new Error(
-          `${NX_PREFIX} The {workspaceRoot} token is only valid at the beginning of an option. (${key})`
-        );
-      }
-      value = value.replace(/\{projectRoot\}/g, project.root);
-      result[opt] = value.replace(/\{projectName\}/g, project.name);
-    } else if (typeof value === 'object' && value) {
-      result[opt] = resolveNxTokensInOptions(
-        value,
-        project,
-        [key, opt].join('.')
-      );
-    }
-  }
-  return result;
-}
-
-export function readTargetDefaultsForTarget(
-  targetName: string,
-  targetDefaults: TargetDefaults,
-  executor?: string
-): TargetDefaults[string] {
-  if (executor && targetDefaults?.[executor]) {
-    // If an executor is defined in project.json, defaults should be read
-    // from the most specific key that matches that executor.
-    // e.g. If executor === run-commands, and the target is named build:
-    // Use, use nx:run-commands if it is present
-    // If not, use build if it is present.
-    return targetDefaults?.[executor];
-  } else if (targetDefaults?.[targetName]) {
-    // If the executor is not defined, the only key we have is the target name.
-    return targetDefaults?.[targetName];
-  }
-
-  let matchingTargetDefaultKey: string | null = null;
-  for (const key in targetDefaults ?? {}) {
-    if (isGlobPattern(key) && minimatch(targetName, key)) {
-      if (
-        !matchingTargetDefaultKey ||
-        matchingTargetDefaultKey.length < key.length
-      ) {
-        matchingTargetDefaultKey = key;
-      }
-    }
-  }
-  if (matchingTargetDefaultKey) {
-    return targetDefaults[matchingTargetDefaultKey];
-  }
-
-  return null;
-}
-
-function createRootMap(projectRootMap: Record<string, ProjectConfiguration>) {
-  const map: Record<string, string> = {};
-  for (const projectRoot in projectRootMap) {
-    const projectName = projectRootMap[projectRoot].name;
-    map[projectRoot] = projectName;
-  }
-  return map;
-}
-
-function resolveCommandSyntacticSugar(
-  target: TargetConfiguration,
-  key: string
-): TargetConfiguration {
-  const { command, ...config } = target ?? {};
-
-  if (!command) {
-    return target;
-  }
-
-  if (config.executor) {
-    throw new Error(
-      `${NX_PREFIX} Project at ${key} should not have executor and command both configured.`
-    );
-  } else {
-    return {
-      ...config,
-      executor: 'nx:run-commands',
-      options: {
-        ...config.options,
-        command: command,
-      },
-    };
-  }
-}
-
-/**
- * Expand's `command` syntactic sugar, replaces tokens in options, and adds information from executor schema.
- * @param target The target to normalize
- * @param project The project that the target belongs to
- * @returns The normalized target configuration
- */
-export function normalizeTarget(
-  target: TargetConfiguration,
-  project: ProjectConfiguration,
-  workspaceRoot: string,
-  projectsMap: Record<string, ProjectConfiguration>,
-  errorMsgKey: string
-) {
-  target = {
-    ...target,
-    configurations: {
-      ...target.configurations,
-    },
-  };
-
-  target = resolveCommandSyntacticSugar(target, project.root);
-
-  target.options = resolveNxTokensInOptions(
-    target.options,
-    project,
-    errorMsgKey
-  );
-
-  for (const configuration in target.configurations) {
-    target.configurations[configuration] = resolveNxTokensInOptions(
-      target.configurations[configuration],
-      project,
-      `${project.root}:${target}:${configuration}`
-    );
-  }
-
-  target.parallelism ??= true;
-
-  if (target.executor && !('continuous' in target)) {
-    try {
-      const [executorNodeModule, executorName] = parseExecutor(target.executor);
-
-      const { schema } = getExecutorInformation(
-        executorNodeModule,
-        executorName,
-        workspaceRoot,
-        projectsMap
-      );
-
-      if (schema.continuous) {
-        target.continuous ??= schema.continuous;
-      }
-    } catch (e) {
-      // If the executor is not found, we assume that it is not a valid executor.
-      // This means that we should not set the continuous property.
-      // We could throw an error here, but it would be better to just ignore it.
-    }
-  }
-
-  return target;
 }
